finalize consumer and producer pattern with shared buffer

+++ src/ProducerConsumerPattern/Operation.java
@@ -4,10 +4,10 @@ package ProducerConsumerPattern;
  * 一个算术运算：两个操作数 + 一个运算符。
  */
 public class Operation {
-    private int v1;
-    private int v2;
+    private double v1;
+    private double v2;
     private char op;
-    public Operation(int v1, int v2, char op){
+    public Operation(double v1, double v2, char op){
 
         this.op = op;
         this.v1 = v1;
@@ -18,11 +18,11 @@ public class Operation {
         return op;
     }
 
-    public int getV2() {
+    public double getV2() {
         return v2;
     }
 
-    public int getV1() {
+    public double getV1() {
         return v1;
     }
     @Override
